refactor of unit tests and now we check for key width at runtime

+++ src/test/peregrine/io/TestFullOuterJoin.java
@@ -86,7 +86,7 @@ public class TestFullOuterJoin extends peregrine.BaseTestWithTwoPartitions {
     }
 
     public static void write( PartitionWriter writer,
-                              int v ) throws IOException {
+                              long v ) throws IOException {
 
     	StructReader key = StructReaders.wrap(v);
     	StructReader value = key;
